Issues fixed

+++ Assets/SimpleHealthBar/Healthbar.cs
@@ -125,7 +125,7 @@ public class Healthbar : MonoBehaviour {
         UpdateHealth();
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(double amount)
     {
         // Remove 'amount' hitpoints, then update the characters health:
         health -= float.Parse(amount.ToString());
